Rename Term Logging Service
Rename the term logging service to reflect that it logs other
things now.
